Display warning about uninstalled plugins in the Administration section of SonarQube
