+++ src/test/java/com/devonfw/ide/sonarqube/common/impl/DevonfwJavaProfileTest.java
@@ -1,7 +1,5 @@
 package com.devonfw.ide.sonarqube.common.impl;
 
-import java.io.File;
-
 import org.junit.Test;
 import org.sonar.api.server.profile.BuiltInQualityProfilesDefinition;
 import org.sonar.plugins.java.Java;
@@ -13,7 +11,6 @@ import com.devonfw.module.test.common.base.ModuleTest;
  *
  */
 public class DevonfwJavaProfileTest extends ModuleTest {
-
   /**
    * Test of {@link DevonfwJavaProfile}
    */
@@ -21,13 +18,10 @@ public class DevonfwJavaProfileTest extends ModuleTest {
   public void test() {
 
     // Create new test profile
-    File pluginDir = new File("src/test/files/qualityprofile/extensions/plugins");
-    DevonfwJavaProfile profileDef = new DevonfwJavaProfile(pluginDir);
-
+    DevonfwJavaProfile profileDef = new DevonfwJavaProfile();
     BuiltInQualityProfilesDefinition.Context context = new BuiltInQualityProfilesDefinition.Context();
     profileDef.define(context);
     BuiltInQualityProfilesDefinition.BuiltInQualityProfile profile = context.profile(Java.KEY, "devonfw Java");
-
     // Assertions
     assertThat(profile.rules()).isNotEmpty();
     assertThat(profile.language()).isEqualTo(Java.KEY);
